[BEAM-6865] refactor InMemoryBagUserStateFactory to standalone class
