+++ runners/flink/src/main/java/org/apache/beam/runners/flink/translation/functions/FlinkExecutableStageFunction.java
@@ -18,13 +18,10 @@
 package org.apache.beam.runners.flink.translation.functions;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.EnumMap;
 import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.function.BiConsumer;
@@ -34,13 +31,9 @@ import org.apache.beam.model.fnexecution.v1.BeamFnApi.ProcessBundleProgressRespo
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.ProcessBundleResponse;
 import org.apache.beam.model.fnexecution.v1.BeamFnApi.StateKey;
 import org.apache.beam.model.pipeline.v1.RunnerApi;
-import org.apache.beam.runners.core.InMemoryStateInternals;
 import org.apache.beam.runners.core.InMemoryTimerInternals;
-import org.apache.beam.runners.core.StateInternals;
 import org.apache.beam.runners.core.StateNamespace;
 import org.apache.beam.runners.core.StateNamespaces;
-import org.apache.beam.runners.core.StateTag;
-import org.apache.beam.runners.core.StateTags;
 import org.apache.beam.runners.core.TimerInternals;
 import org.apache.beam.runners.core.construction.Timer;
 import org.apache.beam.runners.core.construction.graph.ExecutableStage;
@@ -52,6 +45,7 @@ import org.apache.beam.runners.fnexecution.control.ProcessBundleDescriptors;
 import org.apache.beam.runners.fnexecution.control.RemoteBundle;
 import org.apache.beam.runners.fnexecution.control.StageBundleFactory;
 import org.apache.beam.runners.fnexecution.provisioning.JobInfo;
+import org.apache.beam.runners.fnexecution.state.InMemoryBagUserStateFactory;
 import org.apache.beam.runners.fnexecution.state.StateRequestHandler;
 import org.apache.beam.runners.fnexecution.state.StateRequestHandlers;
 import org.apache.beam.runners.fnexecution.translation.BatchSideInputHandlerFactory;
@@ -59,7 +53,6 @@ import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.fn.data.FnDataReceiver;
 import org.apache.beam.sdk.io.FileSystems;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.state.BagState;
 import org.apache.beam.sdk.state.TimeDomain;
 import org.apache.beam.sdk.transforms.join.RawUnionValue;
 import org.apache.beam.sdk.transforms.windowing.BoundedWindow;
@@ -443,95 +436,6 @@ public class FlinkExecutableStageFunction<InputT> extends AbstractRichFunction
           timerDataConsumer.accept(windowedValue, timerData);
         }
       };
-    }
-  }
-
-  /**
-   * Holds user state in memory if the ExecutableStage is stateful. Only one key is active at a time
-   * due to the GroupReduceFunction being called once per key. Needs to be reset via {@code
-   * resetForNewKey()} before processing a new key.
-   */
-  private static class InMemoryBagUserStateFactory
-      implements StateRequestHandlers.BagUserStateHandlerFactory {
-
-    private List<InMemorySingleKeyBagState> handlers;
-
-    private InMemoryBagUserStateFactory() {
-      handlers = new ArrayList<>();
-    }
-
-    @Override
-    public <K, V, W extends BoundedWindow>
-        StateRequestHandlers.BagUserStateHandler<K, V, W> forUserState(
-            String pTransformId,
-            String userStateId,
-            Coder<K> keyCoder,
-            Coder<V> valueCoder,
-            Coder<W> windowCoder) {
-
-      InMemorySingleKeyBagState<K, V, W> bagUserStateHandler =
-          new InMemorySingleKeyBagState<>(userStateId, valueCoder, windowCoder);
-      handlers.add(bagUserStateHandler);
-
-      return bagUserStateHandler;
-    }
-
-    /** Prepares previous emitted state handlers for processing a new key. */
-    void resetForNewKey() {
-      for (InMemorySingleKeyBagState stateBags : handlers) {
-        stateBags.reset();
-      }
-    }
-
-    static class InMemorySingleKeyBagState<K, V, W extends BoundedWindow>
-        implements StateRequestHandlers.BagUserStateHandler<K, V, W> {
-
-      private final StateTag<BagState<V>> stateTag;
-      private final Coder<W> windowCoder;
-
-      /* Lazily initialized state internals upon first access */
-      private volatile StateInternals stateInternals;
-
-      InMemorySingleKeyBagState(String userStateId, Coder<V> valueCoder, Coder<W> windowCoder) {
-        this.windowCoder = windowCoder;
-        this.stateTag = StateTags.bag(userStateId, valueCoder);
-      }
-
-      @Override
-      public Iterable<V> get(K key, W window) {
-        initStateInternals(key);
-        StateNamespace namespace = StateNamespaces.window(windowCoder, window);
-        BagState<V> bagState = stateInternals.state(namespace, stateTag);
-        return bagState.read();
-      }
-
-      @Override
-      public void append(K key, W window, Iterator<V> values) {
-        initStateInternals(key);
-        StateNamespace namespace = StateNamespaces.window(windowCoder, window);
-        BagState<V> bagState = stateInternals.state(namespace, stateTag);
-        while (values.hasNext()) {
-          bagState.add(values.next());
-        }
-      }
-
-      @Override
-      public void clear(K key, W window) {
-        initStateInternals(key);
-        StateNamespace namespace = StateNamespaces.window(windowCoder, window);
-        BagState<V> bagState = stateInternals.state(namespace, stateTag);
-        bagState.clear();
-      }
-
-      private void initStateInternals(K key) {
-        if (stateInternals == null) {
-          stateInternals = InMemoryStateInternals.forKey(key);
-        }
-      }
-
-      void reset() {
-        stateInternals = null;
-      }
     }
   }
 }
